+++ pyinstaller.py
@@ -2,7 +2,7 @@ import os
 
 import PyInstaller.__main__
 
-from src.config import SCRIPT
+from src import SCRIPT
 
 
 def run_pyinstaller(script_path):
