+++ src/interfaces/__init__.py
@@ -1,0 +1,1 @@
+from .baseui import BaseUI
