+++ src/interfaces/graphicalui/__init__.py
@@ -1,0 +1,7 @@
+from .board import Board
+from .button import Button
+from .entry import Entry
+from .label import Label
+from .menu import Menu
+from .menu_handler import MenuHandler
+from .tile import Tile
