+++ src/interfaces/graphicalui/label.py
@@ -2,7 +2,7 @@ from tkinter import ttk
 from tkinter.font import Font
 from typing import List
 
-from src.modules.gameboard.gametile import GameTile
+from src.modules.gameboard import GameTile
 
 
 class Label:
