+++ src/modules/validate/__init__.py
@@ -1,0 +1,3 @@
+from .generate import WordGenerate
+from .validate import WordValidate
+from .wordlist import WordList
