+++ src/utils/__init__.py
@@ -1,0 +1,3 @@
+from .resources import resource_path
+from .timer import Timer
+from .utils import aux_to_indices, is_valid_word, get_letter_point_value
