+++ webapi.py
@@ -1,10 +1,9 @@
 import uvicorn
 from fastapi import FastAPI
 
-from src.config import HOST, PORT, VERSION
-from src.interfaces.baseui import BaseUI
-from src.interfaces.webapi.apirouter import SolverRouter
-from src.interfaces.webapi.baseapi import BaseAPI, BaseRouter
+from src import HOST, PORT, VERSION
+from src.interfaces import BaseUI
+from src.interfaces.webapi import BaseAPI, BaseRouter, SolverRouter
 
 
 class WebAPI(BaseUI):
